scripts/requirements.txt
clickhouse-driver==0.2.7
requests==2.31.0
python-dotenv==1.0.0